MISP to stix, cybox & ciq in 1 module (class methods)
